+++ CHANGELOG.md
@@ -5,7 +5,8 @@
 ### Additions
 
 - Add `getPDAAndBump()` - calculates a Program Derived Address (PDA) and its bump seed from a program address and seeds, automatically encoding different seed types (strings, addresses, and bigints).
-- Add TSDoc comments for all functions.
+- `getTokenAccountBalance()` - can now take either a wallet and token mint (it will find the token account and then get the balance), or a token account address.
+- Add TSDoc comments for all functions, so VSCode and other editors can display parameters nicely.
 - Solana `@solana/kit` has been renamed to `@solana/kit`, and dependencies have been updated accordingly.
 
 ### Bug fixes